+++ R/bestNormalize.R
@@ -1,48 +1,51 @@
 #' Calculate and perform best normalizing transformation
-#' 
+#'
 #' @aliases predict.bestNormalize
-#'   
+#'
 #' @description Performs a suite of normalizing transformations, and selects the
-#'   best one on the basis of the Pearson P test statistic for normality. The 
-#'   transformation that has the lowest P (calculated on the transformed data) 
+#'   best one on the basis of the Pearson P test statistic for normality. The
+#'   transformation that has the lowest P (calculated on the transformed data)
 #'   is selected. See details for more information.
-#'   
+#'
 #' @param x A vector to normalize
 #' @param allow_orderNorm set to FALSE if orderNorm should not be applied
+#' @param allow_lambert Set to TRUE if lambertW should be applied (see details)
 #' @param standardize If TRUE, the transformed values are also centered and
-#'   scaled, such that the transformation attempts a standard normal. This
-#'   will not change the normality statistic.
+#'   scaled, such that the transformation attempts a standard normal. This will
+#'   not change the normality statistic.
 #' @param out_of_sample if FALSE, estimates quickly in-sample performance
 #' @param cluster name of cluster set using \code{makeCluster}
 #' @param k number of folds
 #' @param r number of repeats
+#' @param warn Should bestNormalize warn when a method doesn't work?
 #' @param newdata a vector of data to be (reverse) transformed
 #' @param inverse if TRUE, performs reverse transformation
 #' @param object an object of class 'bestNormalize'
 #' @param ... additional arguments
-#' @details \code{bestNormalize} estimates the optimal normalizing 
-#'   transformation. This transformation can be performed on new data, and 
+#' @details \code{bestNormalize} estimates the optimal normalizing
+#'   transformation. This transformation can be performed on new data, and
 #'   inverted, via the \code{predict} function.
-#'   
-#' @return A list of class \code{bestNormalize} with elements 
-#' 
-#' \item{x.t}{transformed original data} 
-#' \item{x}{original data} 
-#' \item{norm_stats}{Pearson's Pearson's P / degrees of freedom}
-#' \item{method}{out-of-sample or in-sample, number of folds + repeats}
-#' \item{chosen_transform}{the chosen transformation (of appropriate class)}
-#' \item{other_transforms}{the other transformations (of appropriate class)}
-#' 
-#' The \code{predict} function returns the numeric value of the transformation
-#' performed on new data, and allows for the inverse transformation as well.
-#' 
+#'
+#' @return A list of class \code{bestNormalize} with elements
+#'
+#'   \item{x.t}{transformed original data} \item{x}{original data}
+#'   \item{norm_stats}{Pearson's Pearson's P / degrees of freedom}
+#'   \item{method}{out-of-sample or in-sample, number of folds + repeats}
+#'   \item{chosen_transform}{the chosen transformation (of appropriate class)}
+#'   \item{other_transforms}{the other transformations (of appropriate class)}
+#'
+#'   The \code{predict} function returns the numeric value of the transformation
+#'   performed on new data, and allows for the inverse transformation as well.
+#'
 #' @details This function currently estimates the Yeo-Johnson transformation,
-#'   the Box Cox transformation (if the data is positive), and the Lambert WxF
-#'   Gaussianizing transformation of type "s". If allow_orderNorm == TRUE and if
-#'   out_of_sample == FALSE then the ordered quantile normalization technique
-#'   will likely be chosen since it essentially forces the data to follow a
-#'   normal distribution. More information on the orderNorm technique can be
-#'   found in the package vignette, or using \code{?orderNorm}.
+#'   the Box Cox transformation (if the data is positive), the log_10(x+a)
+#'   transformation, the square-root (x+a) transformation, and the arcsinh
+#'   transformation. a is set to max(0, -min(x) + eps) by default.  If
+#'   allow_orderNorm == TRUE and if out_of_sample == FALSE then the ordered
+#'   quantile normalization technique will likely be chosen since it essentially
+#'   forces the data to follow a normal distribution. More information on the
+#'   orderNorm technique can be found in the package vignette, or using
+#'   \code{?orderNorm}.
 #'
 #'   Repeated cross-validation is used to estimate the out-of-sample performance
 #'   of each transformation if out_of_sample = TRUE. While this can take some
@@ -59,50 +62,62 @@
 #'   in \code{bestNormalize()}. Use type = "h" or type = 'hh' at risk of not
 #'   having this estimate 1-1 transform. These alternative types are effective
 #'   when the data has exceptionally heavy tails, e.g. the Cauchy distribution.
-#' 
-#' @examples 
-#' 
+#'   Additionally, as of v. 1.2.0, Lambert of type "s" is not used in
+#'   \code{bestNormalize()} since it uses multiple threads on some Linux systems,
+#'   which is not allowed on CRAN checks. Set allow_lambert = TRUE in order to 
+#'   test this transformation as well.
+#'
+#' @examples
+#'
 #' x <- rgamma(100, 1, 1)
-#' 
+#'
 #' \dontrun{
 #' # With Repeated CV
-#' BN_obj <- bestNormalize(x) 
+#' BN_obj <- bestNormalize(x)
 #' BN_obj
 #' p <- predict(BN_obj)
 #' x2 <- predict(BN_obj, newdata = p, inverse = TRUE)
-#' 
+#'
 #' all.equal(x2, x)
 #' }
-#' 
+#'
 #' # Without Repeated CV
-#' BN_obj <- bestNormalize(x, allow_orderNorm = FALSE, out_of_sample = FALSE) 
+#' BN_obj <- bestNormalize(x, allow_orderNorm = FALSE, out_of_sample = FALSE)
 #' BN_obj
 #' p <- predict(BN_obj)
 #' x2 <- predict(BN_obj, newdata = p, inverse = TRUE)
-#' 
+#'
 #' all.equal(x2, x)
-#' 
-#' 
-#' @seealso  \code{\link[bestNormalize]{boxcox}},
-#'  \code{\link{lambert}}, 
-#'  \code{\link{orderNorm}},
-#'  \code{\link{yeojohnson}} 
+#'
+#'
+#' @seealso  \code{\link[bestNormalize]{boxcox}}, \code{\link{orderNorm}},
+#'   \code{\link{yeojohnson}}
 #' @export
 bestNormalize <- function(x, standardize = TRUE, allow_orderNorm = TRUE,
+                          allow_lambert = FALSE,
                           out_of_sample = TRUE, cluster = NULL, k = 10, 
-                          r = 5) {
+                          warn = TRUE, r = 5) {
   stopifnot(is.numeric(x))
   x.t <- list()
-  methods <- c('lambert', 'yeojohnson', 'boxcox')
+  methods <- c("arcsinh_x", 'boxcox', "exp_x", "log_x", "sqrt_x", 'yeojohnson')
   if (allow_orderNorm) 
     methods <- c(methods, 'orderNorm')
+  if(allow_lambert)
+    methods <- c(methods, "lambert")
   
+  methods <- methods[order(methods)]
+    
   for(i in methods) {
-    trans_i <- try(do.call(i, list(x = x, standardize = standardize)), silent = TRUE)
-    if(is.character(trans_i))
-      warning(paste(i, ' did not work; ', trans_i))
-    else
-      x.t[[i]] <- trans_i
+    args <- list(x=x, standardize = standardize)
+    if(i %in% c("orderNorm", "exp_x", "log_x")) {
+      args[['warn']] <- warn
+    }
+  
+    trans_i <- try(do.call(i, args), silent = TRUE)
+    if(is.character(trans_i)) {
+      if(warn) 
+        warning(paste(i, ' did not work; ', trans_i))
+    } else x.t[[i]] <- trans_i
   }
 
   # Select methods that worked
@@ -112,7 +127,7 @@ bestNormalize <- function(x, standardize = TRUE, allow_orderNorm = TRUE,
   if(out_of_sample) {
     k <- as.integer(k)
     r <- as.integer(r)
-    reps <- get_oos_estimates(x, methods, k, r, cluster)
+    reps <- get_oos_estimates(x, standardize, methods, k, r, cluster)
     norm_stats <- colMeans(reps)
     best_idx <- names(which.min(norm_stats))
     method <- paste("Out-of-sample via CV with", k, "folds and", r, "repeats")
@@ -152,13 +167,21 @@ predict.bestNormalize <- function(object, newdata = NULL, inverse = FALSE, ...) 
 print.bestNormalize <- function(x, ...) {
   prettyD <- paste0(
     'Estimated Normality Statistics (Pearson P / df, lower => more normal):\n',
-    ifelse(length(x$norm_stats['boxcox']), 
+    ifelse("boxcox" %in% names(x$norm_stats), 
            paste(" - Box-Cox:", round(x$norm_stats['boxcox'], 4), '\n'), ''),
-    ifelse(length(x$norm_stats['lambert']), 
+    ifelse("lambert" %in% names(x$norm_stats), 
            paste(" - Lambert's W:", round(x$norm_stats['lambert'], 4), '\n'), ''),
-    ifelse(length(x$norm_stats['yeojohnson']), 
+    ifelse("log_x" %in% names(x$norm_stats), 
+           paste(" - Log_b(x+a):", round(x$norm_stats['log_x'], 4), '\n'), ''),
+    ifelse("sqrt_x" %in% names(x$norm_stats), 
+           paste(" - sqrt(x+a):", round(x$norm_stats['sqrt_x'], 4), '\n'), ''),
+    ifelse("exp_x" %in% names(x$norm_stats), 
+           paste(" - exp(x):", round(x$norm_stats['exp_x'], 4), '\n'), ''),
+    ifelse("arcsinh_x" %in% names(x$norm_stats), 
+           paste(" - arcsinh(x):", round(x$norm_stats['arcsinh_x'], 4), '\n'), ''),
+    ifelse("yeojohnson" %in% names(x$norm_stats), 
            paste(" - Yeo-Johnson:", round(x$norm_stats['yeojohnson'], 4), '\n'), ''),
-    ifelse(length(x$norm_stats['orderNorm']), 
+    ifelse("orderNorm" %in% names(x$norm_stats), 
            paste(" - orderNorm:", round(x$norm_stats['orderNorm'], 4), '\n'), ''), 
     "Estimation method: ", x$method, '\n')
   
@@ -167,7 +190,7 @@ print.bestNormalize <- function(x, ...) {
   print(x$chosen_transform)
 }
 
-get_oos_estimates <- function(x, norm_methods, k, r, cluster) {
+get_oos_estimates <- function(x, standardize, norm_methods, k, r, cluster) {
   x <- x[!is.na(x)]
   fold_size <- floor(length(x) / k)
   if(fold_size < 20) warning("fold_size is ", fold_size, " (< 20), therefore P/df estimates may be off")